+++ R/data.R
@@ -12,6 +12,7 @@
 #' \item{max_year}{The integer representing the maximum year to be analyzed.}
 #' \item{iea_data_path}{A string representing the path to the file containing IEA extended energy balance data.}
 #' \item{ceda_data_folder}{A string representing the path to the file containing CEDA data.}
+#' \item{machine_data_path}{A string representing the path to the folder containing machine data excel files.}
 #' \item{exemplar_table_path}{A string representing the path to the exemplar table.}
 #' \item{fu_analysis_folder}{A string representing the folder containing final-to-useful analyses for various countries.}
 #' \item{report_source_folders}{A string vector representing folders containing reports to be run as the last step of the workflow.}