+++ R/plans.R
@@ -12,6 +12,7 @@
 #' * `max_year`: The maximum year to be analyzed, supplied in the `max_year` argument.
 #' * `iea_data_path`: The path to IEA extended energy balance data, supplied in the `iea_data_path` argument.
 #' * `ceda_data_folder`: The path to the CEDA data, supplied in the `ceda_data_folder` argument.
+#' * `machine_data_path`: The path to the machine data excel files, supplied in the `machine_data_path` argument.
 #' * `exemplar_table_path`: The path to an exemplar table, supplied in the `exemplar_table_path` argument.
 #' * `fu_analysis_folder`: The path to the final-to-useful analysis folder, supplied in the `fu_analysis_folder` argument.
 #' * `report_output_folder`: The path to a report output folder, supplied in the `report_output_folder` argument.
@@ -71,6 +72,7 @@
 #'                Default is "all_targets" to indicate all targets of the plan should be returned.
 #' @param iea_data_path The path to IEA extended energy balance data in .csv format.
 #' @param ceda_data_folder The path to the CEDA data in text file, .per, format.
+#' @param machine_data_path The path to the machine data in .xlsx format.
 #' @param exemplar_table_path The path to an exemplar table.
 #' @param fu_analysis_folder The path to a folder containing final-to-useful analyses.
 #'                           Sub-folders named with 3-letter country abbreviations are assumed.
@@ -94,13 +96,15 @@
 #'          max_year = 1999,
 #'          iea_data_path = "iea_path",
 #'          ceda_data_folder = "ceda_path",
+#'          machine_data_path = "machine_path",
 #'          exemplar_table_path = "exemplar_path",
 #'          fu_analysis_folder = "fu_folder",
 #'          reports_source_folders = "reports_source_folders",
 #'          reports_dest_folder = "reports_dest_folder")
 get_plan <- function(countries, additional_exemplar_countries = NULL,
                      max_year, how_far = "all_targets",
-                     iea_data_path, ceda_data_folder, exemplar_table_path,
+                     iea_data_path, ceda_data_folder,
+                     machine_data_path, exemplar_table_path,
                      fu_analysis_folder, reports_source_folders, reports_dest_folder) {
 
   # Get around warnings of type "no visible binding for global variable".
@@ -135,6 +139,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
     max_year = !!max_year,
     iea_data_path = !!iea_data_path,
     ceda_data_folder = !!ceda_data_folder,
+    machine_data_path = !!machine_data_path,
     exemplar_table_path = !!exemplar_table_path,
     fu_analysis_folder = !!fu_analysis_folder,
     reports_source_folders = !!reports_source_folders,
@@ -154,7 +159,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
                                                              agg_cru_cy_year = 2020)),
     # (1c) Grab Machine data for ALL countries
 
-    AllMachineData = drake::target(read_all_eta_files(eta_fin_paths = get_eta_filepaths())),
+    AllMachineData = drake::target(read_all_eta_files(eta_fin_paths = get_eta_filepaths(machine_data_path))),
     MachineData = drake::target(AllMachineData), # Need to check with Matt how to filter out/extract for countries in countries.