+++ NEWS.md
@@ -1,3 +1,5 @@
+# SEAPSUTWorkflow 0.2.0 (2021-01-23)
+
 * New target "AllMachineData" which consists of a tidy data 
   frames containing information obtained by calling functions in 
   machine_functions.R.